Use same api for get/set values
