+++ core/mapper/exprmapper/json/get.go
@@ -3,7 +3,6 @@ package json
 import (
 	"fmt"
 	"github.com/TIBCOSoftware/flogo-lib/core/mapper/exprmapper/json/field"
-	"reflect"
 	"strconv"
 	"strings"
 
@@ -28,7 +27,7 @@ func ResolvePathValue(value interface{}, refPath string) (interface{}, error) {
 		}
 		return value, nil
 	}
-	return GetFieldValueFromIn(value, mappingField)
+	return GetFieldValue(value, mappingField)
 }
 
 func toInterface(value interface{}) (interface{}, error) {
@@ -52,12 +51,12 @@ func toInterface(value interface{}) (interface{}, error) {
 	return paramMap, nil
 }
 
-func GetFieldValueFromIn(data interface{}, mappingField *field.MappingField) (interface{}, error) {
+func GetFieldValue(data interface{}, mappingField *field.MappingField) (interface{}, error) {
 	var jsonParsed *Container
 	var err error
-
-	if reflect.TypeOf(data).Kind() == reflect.String {
-		jsonParsed, err = ParseJSON([]byte(data.(string)))
+	value, ok := data.(string)
+	if ok {
+		jsonParsed, err = ParseJSON([]byte(value))
 	} else {
 		b, err := json.Marshal(data)
 		if err != nil {
@@ -74,101 +73,7 @@ func GetFieldValueFromIn(data interface{}, mappingField *field.MappingField) (in
 }
 
 func getFieldValue(jsonData *JSONData, mappingField *field.MappingField) (interface{}, error) {
-	//var res interface{}
-	//var err error
-
 	return handleGetSpecialFields(jsonData, mappingField.Getfields())
-
-	//if mappingField.HasArray() && mappingField.HasSepcialField() {
-	//	res, err = handleGetSpecialFields(jsonData, mappingField.Getfields())
-	//} else if mappingField.HasArray() {
-	//	//data, err := getFiledContainer(jsonData, strings.Join(mappingField.Getfields(), "."))
-	//	//if data != nil {
-	//	//	return data.Data(), err
-	//	//}
-	//	res, err = handleGetSpecialFields(jsonData, mappingField.Getfields())
-	//	return nil, err
-	//} else if mappingField.HasSepcialField() {
-	//	res, err = handleGetSpecialFields(jsonData, mappingField.Getfields())
-	//} else {
-	//
-	//	data, err := getFiledContainer(jsonData, strings.Join(mappingField.Getfields(), "."))
-	//	return data.Data(), err
-	//}
-	//return res, err
-}
-
-//func getFieldValueP(jsonData *JSONData, path string) (interface{}, error) {
-//	var res interface{}
-//	var err error
-//	if field.HasArray() && field.HasSpecialFields(path) {
-//		fields, errs := field.GetAllspecialFields(path)
-//		if errs != nil {
-//			return nil, errs
-//		}
-//		res, err = handleGetSpecialFields(jsonData, fields)
-//	} else if field.HasArray(path) {
-//		data, err := getFiledContainer(jsonData, path)
-//		return data.Data(), err
-//	} else if field.HasSpecialFields(path) {
-//		fields, errs := field.GetAllspecialFields(path)
-//		if errs != nil {
-//			return nil, errs
-//		}
-//		res, err = handleGetSpecialFields(jsonData, fields)
-//	} else {
-//		data, err := getFiledContainer(jsonData, path)
-//		return data.Data(), err
-//	}
-//
-//	return res, err
-//}
-
-func getFiledContainer(jsonData *JSONData, path string) (*Container, error) {
-	jsonData.rw.RLock()
-
-	defer jsonData.rw.RUnlock()
-	container := jsonData.container
-	if strings.Index(path, "[") >= 0 {
-		//Array
-
-		arrayFieldName := getFieldName(path)
-		index, err := getFieldSliceIndex(path)
-		if err != nil {
-			return nil, fmt.Errorf("Get array index error, due to [%s]", err.Error())
-		}
-
-		var jsonField *Container
-		if arrayFieldName == "" {
-			//Root level array handling
-			jsonField, err = container.ArrayElement(index)
-		} else {
-			log.Debug("Field Name:", arrayFieldName, " and index: ", index)
-			jsonField, err = container.ArrayElementP(index, arrayFieldName)
-
-		}
-		if err != nil {
-			return nil, err
-		}
-
-		restPath := getRestArrayFieldName(path)
-		if restPath == "" {
-			// value := jsonField.Data()
-			// log.Debug("Type :", reflect.TypeOf(value))
-			// log.Debug("Value :", value)
-			return jsonField, nil
-		}
-		if strings.Index(restPath, "[") > 0 {
-			return getFiledContainer(&JSONData{container: jsonField, rw: sync.RWMutex{}}, restPath)
-		}
-		filed := jsonField.Path(restPath)
-		return filed, nil
-
-	}
-	// value := container.Path(path).Data()
-	// log.Debug("Type :", reflect.TypeOf(value))
-	// log.Debug("Value :", value)
-	return container.Path(path), nil
 }
 
 func getRestArrayFieldName(fieldName string) string {
