+++ core/mapper/exprmapper/json/get_test.go
@@ -1,6 +1,8 @@
 package json
 
 import (
+	"encoding/json"
+	"strconv"
 	"sync"
 	"testing"
 
@@ -54,21 +56,21 @@ var jsonData = `{
 
 func TestRootChildArray(t *testing.T) {
 	mappingField := field.NewMappingField(false, true, []string{"City[0]"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 }
 
 func TestRoot(t *testing.T) {
 	mappingField := field.NewMappingField(false, false, []string{"City"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 }
 
 func TestGetFieldWithSpaces(t *testing.T) {
 	mappingField := field.NewMappingField(false, false, []string{"hello world"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, "CHINA", value)
@@ -76,7 +78,7 @@ func TestGetFieldWithSpaces(t *testing.T) {
 
 func TestGetFieldWithTag(t *testing.T) {
 	mappingField := field.NewMappingField(false, false, []string{"tag  world"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, "CHINA", value)
@@ -84,7 +86,7 @@ func TestGetFieldWithTag(t *testing.T) {
 
 func TestGetArray(t *testing.T) {
 	mappingField := field.NewMappingField(false, true, []string{"Emails[0]"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, "123@123.com", value)
@@ -93,7 +95,7 @@ func TestGetArray(t *testing.T) {
 
 func TestMultipleLevelArray(t *testing.T) {
 	mappingField := field.NewMappingField(false, true, []string{"City[0]", "Array[1]", "id"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, "2222", value)
@@ -101,7 +103,7 @@ func TestMultipleLevelArray(t *testing.T) {
 
 func TestMultipleLevelObject(t *testing.T) {
 	mappingField := field.NewMappingField(false, true, []string{"City[0]", "Park", "Maps", "bb"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, "bb", value)
@@ -109,7 +111,7 @@ func TestMultipleLevelObject(t *testing.T) {
 
 func TestID(t *testing.T) {
 	mappingField := field.NewMappingField(false, false, []string{"Id"})
-	value, err := GetFieldValueFromIn(jsonData, mappingField)
+	value, err := GetFieldValue(jsonData, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, float64(1234), value)
@@ -171,7 +173,7 @@ func TestGetFieldValue(t *testing.T) {
 }
 `
 	mappingField := field.NewMappingField(false, true, []string{"Account", "records[0]", "Name"})
-	value, err := GetFieldValueFromIn(account, mappingField)
+	value, err := GetFieldValue(account, mappingField)
 	log.Infof("Value:%s", value)
 
 	assert.Nil(t, err)
@@ -195,7 +197,7 @@ func TestConcurrentGetk(t *testing.T) {
 				}
 			}()
 			mappingField := field.NewMappingField(false, true, []string{"City[0]", "Park", "Maps", "bb"})
-			value, err := GetFieldValueFromIn(jsonData, mappingField)
+			value, err := GetFieldValue(jsonData, mappingField)
 			assert.Nil(t, err)
 			assert.NotNil(t, value)
 		}(r)
@@ -220,7 +222,7 @@ func TestRootArray(t *testing.T) {
 ]`
 
 	mappingField := field.NewMappingField(false, true, []string{"[0]", "MessageId"})
-	value, err := GetFieldValueFromIn(jsonArray, mappingField)
+	value, err := GetFieldValue(jsonArray, mappingField)
 	assert.Nil(t, err)
 	assert.NotNil(t, value)
 	assert.Equal(t, "1c0483d9-8166-4df0-be9f-cd03177a38c6", value)
@@ -252,7 +254,7 @@ func TestRootArrayInvalid(t *testing.T) {
 ]`
 
 	mappingField := field.NewMappingField(false, true, []string{"[0]", "MessageId[0]"})
-	value, err := GetFieldValueFromIn(jsonArray, mappingField)
+	value, err := GetFieldValue(jsonArray, mappingField)
 	assert.NotNil(t, err)
 	assert.Nil(t, nil, value)
 
@@ -273,7 +275,7 @@ func TestGetValue(t *testing.T) {
 
 	mappingField := field.NewMappingField(false, false, []string{"id"})
 
-	v, err := GetFieldValueFromIn(value, mappingField)
+	v, err := GetFieldValue(value, mappingField)
 	assert.Nil(t, err)
 	assert.Equal(t, "12222", v)
 
@@ -286,28 +288,216 @@ func TestGetValue(t *testing.T) {
 	testMap2["id2"] = int(2)
 
 	mappingField2 := field.NewMappingField(false, false, []string{"id"})
-	v, err = GetFieldValueFromIn(testMap, mappingField2)
+	v, err = GetFieldValue(testMap, mappingField2)
 	assert.Nil(t, err)
 	assert.Equal(t, "id", v)
 
 	mappingField3 := field.NewMappingField(false, false, []string{"id2"})
-	v, err = GetFieldValueFromIn(testMap2, mappingField3)
+	v, err = GetFieldValue(testMap2, mappingField3)
 	assert.Nil(t, err)
 	assert.Equal(t, float64(2), v)
 
 	mappingField4 := field.NewMappingField(false, false, []string{"id", "id"})
-	v, err = GetFieldValueFromIn(testMap2, mappingField4)
+	v, err = GetFieldValue(testMap2, mappingField4)
 	assert.Nil(t, err)
 	assert.Equal(t, "12222", v)
 
 	////Int64
 	mappingFieldInt64 := field.NewMappingField(false, false, []string{"id", "int_64"})
-	v, err = GetFieldValueFromIn(testMap2, mappingFieldInt64)
+	v, err = GetFieldValue(testMap2, mappingFieldInt64)
 	assert.Nil(t, err)
 	assert.Equal(t, float64(123), v)
 	//Int
 	mappingFieldint := field.NewMappingField(false, false, []string{"id", "int_v"})
-	v, err = GetFieldValueFromIn(testMap2, mappingFieldint)
+	v, err = GetFieldValue(testMap2, mappingFieldint)
 	assert.Nil(t, err)
 	assert.Equal(t, float64(12), v)
 }
+
+
+var SpecialData = `{
+    "City": [
+        {
+            "Array": [
+                {
+                    "id": "11111"
+                },
+                {
+                    "id": "2222"
+                }
+            ],
+            "InUS": true,
+            "Name": "Sugar Land",
+            "Park": {
+                "Emails": null,
+                "Location": "location",
+                "Maps": {
+                    "bb": "bb",
+                    "cc": "cc",
+                    "dd": "dd"
+                },
+                "Name": "Name"
+            }
+        }
+    ],
+    "Emails": [
+        "123@123.com",
+        "456@456.com"
+    ],
+    "Id": 1234,
+    "In": "string222",
+    "Maps": {
+        "bb.bb": {
+            "id": "10001"
+        },
+        "cc#cc": "cc",
+        "dd**cc": "dd"
+    },
+    "Maps2": {
+        "bb.bb": [
+            {
+                "id": "10001"
+            }
+        ],
+        "good":[{"id":"12", "x.y":"234"}],
+        "cc#cc": "cc",
+        "dd**cc": "dd"
+    },
+    "Maps3": {
+        "bb.bb": [
+            {
+                "x.y": "10001"
+            }
+        ],
+        "cc#cc": [
+            {
+                "x.y": {
+                    "id": "1"
+                }
+            }
+        ],
+        "dd*cc": {
+            "x.y": {
+                "g%f": "hello"
+            },
+            "y.x":[
+                {"d.d":"123"}
+            ]
+        }
+    }
+}`
+
+func TestSpecialFieldNames(t *testing.T) {
+	mappingField := field.NewMappingField(true, false, []string{"Maps", "bb.bb"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	log.Info(value)
+	v, _ := json.Marshal(value)
+	assert.Equal(t, `{"id":"10001"}`, string(v))
+}
+func TestGetSpecialFieldRoot(t *testing.T) {
+	mappingField := field.NewMappingField(true, false, []string{"Maps", "bb.bb"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, map[string]interface{}{"id": "10001"}, value)
+}
+
+func TestGetSpecial2LevelObjectField(t *testing.T) {
+	mappingField := field.NewMappingField(true, false, []string{"Maps", "bb.bb", "id"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, "10001", value)
+}
+
+func TestGetSpecial2LevelArrayField(t *testing.T) {
+	mappingField := field.NewMappingField(true, true, []string{"Maps2", "bb.bb[0]", "id"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, "10001", value)
+}
+
+func TestGetSpecial2LevelArrayField2(t *testing.T) {
+	mappingField := field.NewMappingField(true, true, []string{"Maps3", "cc#cc[0]", "x.y", "id"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, "1", value)
+}
+
+func TestGetSpecialSpecial(t *testing.T) {
+	mappingField := field.NewMappingField(true, false, []string{"Maps3", "dd*cc", "x.y", "g%f"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, "hello", value)
+}
+
+func TestGetSpecialSpecial2(t *testing.T) {
+	mappingField := field.NewMappingField(true, false, []string{"Maps3", "dd*cc", "y.x[0]", "d.d"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, "123", value)
+}
+
+func TestGetSpecialSpecial3(t *testing.T) {
+	mappingField := field.NewMappingField(true, false, []string{"Maps2", "good[0]", "x.y"})
+	value, err := GetFieldValue(SpecialData, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+	assert.Equal(t, "234", value)
+}
+
+func TestHasArrayFieldInArray(t *testing.T) {
+	assert.False(t, hasArrayFieldInArray([]string{"[square]"}))
+	assert.True(t, hasArrayFieldInArray([]string{"bb.bb[0]"}))
+}
+
+func TestSpecial(t *testing.T) {
+
+	data := `{
+        "[square]":"123",
+  "array1": [
+    {
+      "id.1": 21907387
+    },
+    {
+      "email": -54931037,
+      "array2": [
+        {
+          "id.2": 3458316
+        },
+        {
+          "id.2": 57420133
+        },
+        {
+          "id.2": -95395610
+        },
+        {
+          "id.2": 68243245
+        }
+      ]
+    }
+  ]
+}`
+
+	mappingField := field.NewMappingField(true, true, []string{"array1[1]", "array2[0]", "id.2"})
+	value, err := GetFieldValue(data, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+
+	assert.Equal(t, "3458316", strconv.FormatFloat(value.(float64), 'f', -1, 64))
+
+	mappingField = field.NewMappingField(true, false, []string{"[square]"})
+	value, err = GetFieldValue(data, mappingField)
+	assert.Nil(t, err)
+	assert.NotNil(t, value)
+
+	assert.Equal(t, "123", value)
+
+}
+
